Move BasicTableLayout into foundation.layout package
